Move OpenVPN configuration to subview
